get_deletion_logs_for_snapshot
